requirements.txt
pandas>=1.5
numpy>=1.23
scipy>=1.9
scikit-learn>=1.1
umap-learn>=0.5
matplotlib>=3.6
seaborn>=0.12
pytest>=7.0
